pyproject.toml
[tool.poetry]
name = "flc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.2.6"
pandas = "^2.3.0"
pyyaml = "^6.0.2"
tqdm = "^4.67.1"
pydantic = "^2.11.5"
scikit-learn = "^1.7.0"
click = "^8.2.1"
pytest = "^8.4.0"
matplotlib = "^3.10.3"
seaborn = "^0.13.2"
xgboost = "^3.0.2"
umap-learn = "^0.5.8"
plotly = "^6.2.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
tabulate = "^0.9.0"
nbformat = "^5.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
